move MyPlaybackStatusMonitor from Thread to Service.getMainLooper()
The purpose of this monitor is to poll ExoPlayer at an interval
(currently once every 10 seconds) to obtain its playback state.

This playback state is only used for the GET "/scrub" HTTP endpoint;
which is fairly unimportant, imho.

ExoPlayer gives a warning when accessed from a different Thread.
There doesn't seem to be any harm done by doing so.
There are (currently) other areas of code that do this, as well.
If the only negative is logcat clutter,
then this timer is the worst offender.