[wip] allow HTTP endpoint "/volume" to accept a value greater than 1.0
The range of acceptable values is 0.0 to AUDIO_VOLUME_MAX_PERCENT/100.
The value 0.0 is equivalent to 'mute'.
The value 1.0 is unity gain at 100% input volume.
The value 2.0 is amplified  to 200% input volume.

AUDIO_VOLUME_MAX_PERCENT is an integer resource value,
which is currently set to: 300%

Audio amplification requires Android 4.4 or higher.

status:
=======
* app builds, installs, and runs
* no functional difference to v2.0.24
  - audio amplification has no effect