allow HTTP endpoint "/volume" to accept a value greater than 1.0
The range of acceptable values is 0.0 to (AUDIO_VOLUME_MAX_DB_BOOST + 1.0).
The value 0.0 is   0% input volume (equivalent to 'mute').
The value 1.0 is 100% input volume (unity gain).
The value 6.5 is 100% input volume and amplified by 5.5 dB.

AUDIO_VOLUME_MAX_DB_BOOST is an integer resource value,
which is currently set to: 50 dB

Audio amplification requires Android 4.4 or higher.

housekeeping:
=============
commit is a squashed merge from branch:
  'dev/001-audio-amplification'