release: '002.00.47-16API'
Extend a class in JmDNS to customize the "host" attribute of the discoverable service object,
to change its value from "localhost.local" to the WiFi IP address of ExoAirPlayer (on the LAN).

The motivation for this change is to gain compatibility with "WebTorrent Desktop":
  https://github.com/webtorrent/webtorrent-desktop

Compatibility means that "WebTorrent Desktop" can detect instances of ExoAirPlayer (on the LAN),
and "cast" URLs to a locally hosted proxy server for videos that can be downloaded from the p2p network.