release: '002.00.50-16API'
Add code to properly handle unstable WiFi network conditions.

In older versions of Android, an open/bound ServerSocket
could survive network disconnect followed by subsequent reconnect.

In newer versions of Android, an open/bound ServerSocket
throws a "Socket closed" IOException upon network disconnect.

This added code will detect this condition,
and poll every 15 seconds to determine when the network is available.
At this time, a new ServerSocket is open/bound.
If the IP assigned to the device on the WiFi network is changed,
the foreground service notification is updated.