release: '002.00.52-16API'
Change the methodology used when the app exits
to wait for all downloads to be removed from offline cache
before finally killing the process belonging to the app.

Previously, a fixed-duration timer was used
to create a static period of time during which the task must complete.

Now, the idle state of "DownloadManager" is used as a trigger.
The specifics are described here:
  https://github.com/google/ExoPlayer/issues/10784

Anecdotally, after minimal testing, this solution does appear to work.
The precise methodology might change, should a better solution emerge.