fix API endpoint: "/edit-preferences"
Broken by release: '003.08.00-21API'

Well, not broken..
but the list of recognized keys was invalidated.

In the spirit of DRY (don't repeat yourself)..
some old code was refactored to be reused and shared
with code called when preferences are updated by the Settings activity.
The problem is that the list of keys recognized by the API
is slightly different from the real keys used by SharedPreferences.

Generally speaking, there are 3 solutions:
  1. rename the real keys used by SharedPreferences
     to be identical to the keys recognized by the API..
     which I don't like, and feel like the tail wagging the dog.
  2. change the README documentation and SPA sender for the API endpoint
     to update the list of keys recognized by the API
     to become identical to the real keys used by SharedPreferences..
     which is OK, but breaks userspace.
  3. update the code that is specific to processing preference updates via API,
     to normalize the keys received in the request..
     before they are passed to the shared code that requires the real keys..
     which is the best of both worlds.

In terms of implementing option 3..
a lookup table isn't necessary,
since the only difference between the two sets of key values
is that "_" characters in the real keys were change to "-" in the API keys..
which was originally done so the format of the data entered by the user to configure preferences
is consistent with the format used (by other API endpoints) to specify sets of HTTP request headers.
It's worth noting that the API will now accept keys in either format.